tests individuals created
